refactored find cart items functions
