+++ src/services/productsService.js
@@ -12,7 +12,7 @@ const findAllProductsService = async () => {
   return products;
 };
 
-const findProductsByIds = async (productIds) => {
+const findProductsByIdsService = async (productIds) => {
   const productIdsQuery = ids.map(() => "?").join(", ");
 
   const query = productsServiceQueries.findProductsByIdsQuery(productIdsQuery);
@@ -99,7 +99,7 @@ const decrementProductStockQuantity = async (products) => {
 
 export default {
   findAllProductsService,
-  findProductsByIds,
+  findProductsByIdsService,
   findProductService,
   createProductService,
   updateProductService,
